studentservices/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pip==23.2.1
pytz==2024.1
PyYAML==6.0.1
setuptools==68.2.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
wheel==0.41.2
